Refactored architecture + cmakelists
