+++ odrive_node/CMakeLists.txt
@@ -1,35 +1,55 @@
 cmake_minimum_required(VERSION 3.5)
 project(odrive_node)
 
-# Default to C99
-if(NOT CMAKE_C_STANDARD)
-  set(CMAKE_C_STANDARD 99)
-endif()
-
-# Default to C++14
+# Default to C++17
 if(NOT CMAKE_CXX_STANDARD)
-  set(CMAKE_CXX_STANDARD 14)
+  set(CMAKE_CXX_STANDARD 17)
 endif()
 
 if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
   add_compile_options(-Wall -Wextra -Wpedantic)
 endif()
 
-# find dependencies
+################################################################################
+# Find and load build settings from external packages
+################################################################################
 find_package(ament_cmake REQUIRED)
 find_package(rclcpp REQUIRED)
 find_package(std_msgs REQUIRED)
 
-add_executable(odrive_node src/odrive_node.cpp)
-ament_target_dependencies(odrive_node rclcpp std_msgs)
+################################################################################
+# Build
+################################################################################
+include_directories(
+  include
+)
 
-install(TARGETS
-  odrive_node
-  DESTINATION lib/${PROJECT_NAME})
+set(dependencies
+  "rclcpp"
+  "std_msgs"
+)
 
-if(BUILD_TESTING)
-  find_package(ament_lint_auto REQUIRED)
-  ament_lint_auto_find_test_dependencies()
-endif()
+set(EXEC_NAME "odrive_node")
 
+add_executable(${EXEC_NAME}
+  src/odrive_node.cpp
+  src/ODrive.cpp
+)
+
+ament_target_dependencies(${EXEC_NAME} ${dependencies})
+
+################################################################################
+# Install
+################################################################################
+install(TARGETS ${EXEC_NAME}
+  DESTINATION lib/${PROJECT_NAME}
+)
+
+install(DIRECTORY include/
+  DESTINATION include/
+)
+
+################################################################################
+# Macro for ament package
+################################################################################
 ament_package()
